requirements.txt
alembic==1.4.3
appdirs==1.4.4
astroid==2.5.2
attrs==20.3.0
Authlib==0.15.3
Automat==20.2.0
autopep8==1.5.6
awscli==1.19.39
backports.shutil-get-terminal-size==1.0.0
backports.shutil-which==3.5.2
bcrypt==3.2.0
beautifulsoup4==4.9.3
blinker==1.4
botocore==1.20.39
braceexpand==0.1.6
cachelib==0.1.1
cairocffi==1.2.0
CairoSVG==2.5.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
check50==3.2.2
click==7.1.2
colorama==0.4.3
compare50==1.2.3
constantly==15.1.0
cryptography==3.4.7
cs50==6.0.4
cssselect==1.1.0
cssselect2==0.4.1
cycler==0.10.0
defusedxml==0.7.1
distlib==0.3.1
dnspython==2.0.0
docutils==0.15.2
EditorConfig==0.12.3
email-validator==1.1.2
filelock==3.0.12
flash==1.0.3
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Ext==0.1
Flask-Login==0.5.0
Flask-Session==0.3.2
Flask-SQLAlchemy==1.3.23
Flask-WhooshAlchemy==0.56
Flask-WTF==0.14.3
greenlet==1.0.0
help50==3.0.3
html5lib==1.1
hyperlink==21.0.0
icdiff==1.9.1
idna==3.1
ikp3db==1.4.5
incremental==21.3.0
Inflector==2.0.12
intervaltree==2.1.0
isort==5.8.0
itemadapter==0.2.0
itemloaders==1.0.4
itsdangerous==1.1.0
jellyfish==0.8.2
Jinja2==2.11.3
jmespath==0.10.0
jsbeautifier==1.13.5
kiwisolver==1.3.1
lazy-object-proxy==1.6.0
lib50==2.0.9
lxml==4.6.3
Mako==1.1.3
MarkupSafe==1.1.1
matplotlib==3.4.0
mccabe==0.6.1
natsort==7.1.1
numpy==1.20.0
oauthlib==3.1.0
pandas==1.2.3
parsel==1.6.0
pexpect==4.8.0
Pillow==8.1.2
plotly==4.14.3
Protego==0.1.16
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
PyDispatcher==2.0.5
Pygments==2.8.1
pylint==2.7.4
pylint-django==2.4.2
pylint-flask==0.6
pylint-plugin-utils==0.6
pyOpenSSL==20.0.1
pyparsing==2.4.7
PyPDF2==1.26.0
Pyphen==0.10.0
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
python-magic==0.4.22
pytz==2021.1
PyYAML==5.4.1
queuelib==1.5.0
render50==4.2.1
requests==2.25.1
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.5
s3cmd==2.1.0
s3transfer==0.3.6
Scrapy==2.4.1
service-identity==18.1.0
six==1.15.0
sortedcontainers==2.3.0
soupsieve==2.2.1
SQLAlchemy==1.3.23
sqlparse==0.4.1
style50==2.7.4
submit50==3.0.3
termcolor==1.1.0
tinycss2==1.1.0
toml==0.10.2
tqdm==4.59.0
Twisted==21.2.0
twython==3.8.2
urllib3==1.26.4
virtualenv==20.4.3
w3lib==1.22.0
WeasyPrint==52.4
webencodings==0.5.1
Werkzeug==1.0.1
Whoosh==2.7.4
wrapt==1.12.1
WTForms==2.3.3
zope.interface==5.3.0
